# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VendlySys - SaaS multiempresa para salões e clínicas
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /empresas/{id}:
    get:
      operationId: getEmpresa
      tags: [empresas]
      summary: Busca empresa por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Empresa encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empresa"
    put:
      operationId: updateEmpresa
      tags: [empresas]
      summary: Atualiza empresa
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEmpresaBody"
      responses:
        "200":
          description: Empresa atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empresa"
...
  /clientes:
    get:
      operationId: listClientes
      tags: [clientes]
      summary: Lista clientes da empresa
      parameters:
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Cliente"
    post:
      operationId: createCliente
      tags: [clientes]
      summary: Cria novo cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteBody"
      responses:
        "201":
          description: Cliente criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
  /clientes/{id}:
    get:
      operationId: getCliente
      tags: [clientes]
      summary: Busca cliente por ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    put:
      operationId: updateCliente
      tags: [clientes]
      summary: Atualiza cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClienteBody"
      responses:
        "200":
          description: Cliente atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cliente"
    delete:
      operationId: deleteCliente
      tags: [clientes]
      summary: Remove cliente
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Cliente removido
...
  /despesas:
    get:
      operationId: listDespesas
      tags: [despesas]
      summary: Lista despesas da empresa
      parameters:
        - name: data_inicio
          in: query
          schema:
            type: string
            format: date
        - name: data_fim
          in: query
          schema:
            type: string
            format: date
        - name: categoria
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Lista de despesas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Despesa"
    post:
      operationId: createDespesa
      tags: [despesas]
      summary: Lança nova despesa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDespesaBody"
      responses:
        "201":
          description: Despesa criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Despesa"
  /despesas/{id}:
    put:
      operationId: updateDespesa
      tags: [despesas]
      summary: Atualiza despesa
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDespesaBody"
      responses:
        "200":
          description: Despesa atualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Despesa"
    delete:
      operationId: deleteDespesa
      tags: [despesas]
      summary: Remove despesa
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Despesa removida
...
components:
  schemas:
...
    Empresa:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        slug:
          type: string
        cor_primaria:
          type: string
        cor_secundaria:
          type: string
        logo_url:
          type: string
        telefone:
          type: string
        endereco:
          type: string
      required: [id, nome, slug]
    UpdateEmpresaBody:
      type: object
      properties:
        nome:
          type: string
        cor_primaria:
          type: string
        cor_secundaria:
          type: string
        logo_url:
          type: string
        telefone:
          type: string
        endereco:
          type: string
...
    Cliente:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        telefone:
          type: string
        email:
          type: string
        empresa_id:
          type: integer
        data_nascimento:
          type: string
          format: date
        observacoes:
          type: string
        total_agendamentos:
          type: integer
        total_gasto:
          type: number
      required: [id, nome, empresa_id]
    CreateClienteBody:
      type: object
      properties:
        nome:
          type: string
        telefone:
          type: string
        email:
          type: string
          format: email
        data_nascimento:
          type: string
          format: date
        observacoes:
          type: string
      required: [nome]
...
    Despesa:
      type: object
      properties:
        id:
          type: integer
        empresa_id:
          type: integer
        descricao:
          type: string
        categoria:
          type: string
          enum: [aluguel, produto, funcionario, marketing, equipamento, outros]
        valor:
          type: number
        data:
          type: string
          format: date
        pago:
          type: boolean
        forma_pagamento:
          type: string
          enum: [dinheiro, pix, cartao_credito, cartao_debito, transferencia]
        observacoes:
          type: string
      required: [id, empresa_id, descricao, categoria, valor, data]
    CreateDespesaBody:
      type: object
      properties:
        descricao:
          type: string
        categoria:
          type: string
          enum: [aluguel, produto, funcionario, marketing, equipamento, outros]
        valor:
          type: number
        data:
          type: string
          format: date
        pago:
          type: boolean
        forma_pagamento:
          type: string
          enum: [dinheiro, pix, cartao_credito, cartao_debito, transferencia]
        observacoes:
          type: string
      required: [descricao, categoria, valor, data]
...